src/streamer/requirements.txt
attrs==21.2.0
autobahn==19.11.2
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
EasyProcess==0.3
enum34==1.1.10
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kiteconnect==3.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyVirtualDisplay==2.2
requests==2.25.1
selenium==3.141.0
service-identity==21.1.0
six==1.16.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
xvfbwrapper==0.2.9
zope.interface==5.4.0
